refactor(team): extract cap-compliance logic into TeamCapCalculator
Move getSalaryCapArray, getTotalCurrentSeasonSalaries,
getTotalNextSeasonSalaries, canAddContractWithoutGoingOverHardCap,
canAddBuyoutWithoutExceedingBuyoutLimit (+ the convertPlrResultIntoPlayerArray
helper) off TeamQueryRepository into a new TeamCapCalculator collaborator
(ADR-0028 domain calculator, mirrors FreeAgencyCapCalculator). Repository now
holds only raw row/aggregate fetches. CapSpaceService and ExtensionService
switch to the calculator via constructor injection (nullable default).
