+++ ibl5/classes/Team/TeamQueryRepository.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Team;
 
-use League\League;
-use Player\Player;
 use Team\Contracts\TeamQueryRepositoryInterface;
 use Season\Season;
 use Trading\BuyoutLedgerRepository;
@@ -310,139 +308,5 @@ class TeamQueryRepository extends \BaseMysqliRepository implements TeamQueryRepo
             "i",
             $teamId
         );
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::getSalaryCapArray()
-     *
-     * @return array<string, int>
-     */
-    public function getSalaryCapArray(string $teamName, int $teamId, Season $season): array
-    {
-        /** @var array<string, int> $salaryCapSpent */
-        $salaryCapSpent = [];
-        $resultContracts = $this->getRosterUnderContractOrderedByName($teamId);
-
-        foreach ($resultContracts as $contract) {
-            $yearUnderContract = $contract['cy'] ?? 0;
-            if ($season->isOffseasonPhase()) {
-                $yearUnderContract++;
-            }
-
-            $cyt = $contract['cyt'] ?? 0;
-            $i = 1;
-            while ($yearUnderContract <= $cyt) {
-                $fieldString = "salary_yr" . $yearUnderContract;
-                $key = "year" . $i;
-                if (!isset($salaryCapSpent[$key])) {
-                    $salaryCapSpent[$key] = 0;
-                }
-                $rawSalary = $contract[$fieldString] ?? 0;
-                $salaryCapSpent[$key] += is_numeric($rawSalary) ? (int) $rawSalary : 0;
-                $yearUnderContract++;
-                $i++;
-            }
-        }
-
-        // Add cash considerations (trades, buyouts) for the team
-        $cashRows = $this->cashConsiderationRepo->getTeamCashForSalary($teamId);
-
-        foreach ($cashRows as $cashRow) {
-            $yearUnderContract = $cashRow['cy'];
-            if ($season->isOffseasonPhase()) {
-                $yearUnderContract++;
-            }
-
-            $i = 1;
-            while ($yearUnderContract <= 6) {
-                $key = "year" . $i;
-                if (!isset($salaryCapSpent[$key])) {
-                    $salaryCapSpent[$key] = 0;
-                }
-                $salaryCapSpent[$key] += BuyoutLedgerRepository::salaryForContractYear($cashRow, $yearUnderContract);
-                $yearUnderContract++;
-                $i++;
-            }
-        }
-
-        return $salaryCapSpent;
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::getTotalCurrentSeasonSalaries()
-     *
-     * @param list<PlayerRow> $result
-     */
-    public function getTotalCurrentSeasonSalaries(array $result): int
-    {
-        $totalCurrentSeasonSalaries = 0;
-
-        $playerArray = $this->convertPlrResultIntoPlayerArray($result);
-        foreach ($playerArray as $player) {
-            $totalCurrentSeasonSalaries += $player->getCurrentSeasonSalary();
-        }
-        return $totalCurrentSeasonSalaries;
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::getTotalNextSeasonSalaries()
-     *
-     * @param list<array<string, mixed>> $result
-     */
-    public function getTotalNextSeasonSalaries(array $result): int
-    {
-        $totalNextSeasonSalaries = 0;
-
-        /** @var list<PlayerRow> $typedResult */
-        $typedResult = $result;
-        $playerArray = $this->convertPlrResultIntoPlayerArray($typedResult);
-        foreach ($playerArray as $player) {
-            $totalNextSeasonSalaries += $player->getNextSeasonSalary();
-        }
-        return $totalNextSeasonSalaries;
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::canAddContractWithoutGoingOverHardCap()
-     */
-    public function canAddContractWithoutGoingOverHardCap(int $teamId, int $contractValue): bool
-    {
-        $teamResult = $this->getRosterUnderContractOrderedByName($teamId);
-        $totalCurrentSeasonSalaries = $this->getTotalCurrentSeasonSalaries($teamResult);
-        $projectedTotalCurrentSeasonSalaries = $totalCurrentSeasonSalaries + $contractValue;
-
-        return $projectedTotalCurrentSeasonSalaries <= League::HARD_CAP_MAX;
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::canAddBuyoutWithoutExceedingBuyoutLimit()
-     */
-    public function canAddBuyoutWithoutExceedingBuyoutLimit(int $teamId, int $buyoutValue, Season $season): bool
-    {
-        $buyoutsResult = $this->getBuyouts($teamId);
-        $totalCurrentSeasonBuyouts = BuyoutLedgerRepository::sumCurrentSeasonSalaryFromRows(
-            $buyoutsResult,
-            $season->isOffseasonPhase()
-        );
-        $projectedTotalCurrentSeasonBuyouts = $totalCurrentSeasonBuyouts + $buyoutValue;
-        $buyoutLimit = League::HARD_CAP_MAX * Team::BUYOUT_PERCENTAGE_MAX;
-
-        return $projectedTotalCurrentSeasonBuyouts <= $buyoutLimit;
-    }
-
-    /**
-     * @see TeamQueryRepositoryInterface::convertPlrResultIntoPlayerArray()
-     *
-     * @param list<PlayerRow> $result
-     * @return array<int, Player>
-     */
-    public function convertPlrResultIntoPlayerArray(array $result): array
-    {
-        $array = [];
-        foreach ($result as $plrRow) {
-            $playerID = (int) $plrRow['pid'];
-            $array[$playerID] = Player::withPlrRow($this->db, $plrRow);
-        }
-        return $array;
     }
 }
